Fix circular imports on script invocation
The previous architecture was such that the "web" package was calling the
"data" package, however, a class that should have been defined in "data"
was defined in the "web" package.

This lead to circular imports in certain cases: it did not happen in app
invocations, but only on script calls, which made export impossible.
